requirements.txt
pandas
numpy
scikit-learn
xgboost
mlflow
joblib
fastapi
uvicorn
kafka-python
prometheus_client
python-dotenv
